separate ChainRules-independent code
